% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{diagQF}
\alias{diagQF}
\title{Diagonal Quadratic Form}
\usage{
diagQF(Z, w)
}
\arguments{
\item{Z}{Design matrix.}

\item{w}{Weight vector.}
}
\value{
Numeric matrix.
}
\description{
Calculates the matrix \eqn{Z'WZ}, where \eqn{W} is a diagonal
matrix.
}
