% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{entry_add}
\alias{entry_add}
\alias{entry_subtract}
\title{This adds a number to an entry in matrix}
\usage{
entry_add(M, entry, x)

entry_subtract(M, entry, x)
}
\arguments{
\item{M}{A matrix}

\item{entry}{An entry created using function \link{e}}

\item{x}{A single nunber.}
}
\description{
More details...
}
\examples{
M <- matrix(1:4, 2, 2)
entry_add(M, e(2, 2), 3)

}
