% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{entry_multiply_by}
\alias{entry_multiply_by}
\title{Entry multipled}
\usage{
entry_multiply_by(M, entry, x)
}
\arguments{
\item{M}{A matrix}

\item{entry}{An entry created using function \link{e}}

\item{x}{A single nunber.}
}
\description{
Multiply matrix.... bla bla
}
