% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FitOLS}
\alias{FitOLS}
\title{Ordinary Least Squares}
\usage{
FitOLS(y, X)
}
\arguments{
\item{y}{Nx1 Numeric vector.}

\item{X}{NxP Numeric matrix.}
}
\value{
List of model components.
}
\description{
Fits the standard OLS model.
}
