% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FitWLS}
\alias{FitWLS}
\title{Weighted Least Squares}
\usage{
FitWLS(y, X, w)
}
\arguments{
\item{y}{Nx1 Response vector.}

\item{X}{NxP Design matrix.}

\item{w}{Nx1 Weight vector.}
}
\value{
List of model components.
}
\description{
Fits the following weighted least squares model: 
\eqn{y_{i}=x_{i}'\beta+\epsilon_{i}}. Here, the subject-specific residual is
normally distributed with mean zero and variance \eqn{\sigma^{2}/w_{i}}.
\eqn{w_{i}} is a known, subject-specific weight, and \eqn{\sigma} is a
common scale parameter.
}
