% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{linProj}
\alias{linProj}
\title{Linear Projection}
\usage{
linProj(X, Y)
}
\arguments{
\item{X}{Numeric matrix.}

\item{Y}{Numeric matrix.}
}
\value{
List containing the following:
\item{Coord}{Coordinates of the projection w.r.t. X.}
\item{Para}{Projection onto the image of X.}
\item{Ortho}{Projection onto the orthogonal complement.}
}
\description{
Decomposes matrix \eqn{Y} into the projection onto the image of \eqn{X},
and the projection onto the orthogonal complement of the image.
}
