% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MatCov}
\alias{MatCov}
\title{Covariance}
\usage{
MatCov(A, B, corMat = FALSE)
}
\arguments{
\item{A}{NxP matrix.}

\item{B}{NxQ matrix.}

\item{corMat}{Return correlation matrix?}
}
\value{
Numeric matrix.
}
\description{
Calculates the correlation between two vectors.
}
