% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entries.R
\name{matrix-ops}
\alias{matrix-ops}
\alias{entry_add}
\alias{entry_subtract}
\alias{entry_multiply_by}
\alias{entry_divide_by}
\title{Matrix operations}
\usage{
entry_add(M, entry, x)

entry_subtract(M, entry, x)

entry_multiply_by(M, entry, x)

entry_divide_by(M, entry, x)
}
\arguments{
\item{M}{A matrix.}

\item{entry}{An entry, which is a vector of size.}

\item{x}{The value to add.}
}
\description{
This adds a value to a particular entry of a matrix.
}
\examples{
add_entry(matrix(0, 2, 2), e(1, 1), 3)

}
