% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixops-package.R
\docType{package}
\name{matrixops-package}
\alias{matrixops}
\alias{matrixops-package}
\title{Matrix operations}
\description{
These functions include ....
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/emitanaka/matrixops}
  \item \url{http://emitanaka.org/matrixops/}
  \item Report bugs at \url{https://github.com/emitanaka/matrixops/issues}
}

}
\author{
\strong{Maintainer}: Emi Tanaka \email{dr.emi.tanaka@gmail.com} (\href{https://orcid.org/0000-0002-1455-259X}{ORCID})

}
\keyword{internal}
