% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fitWLS}
\alias{fitWLS}
\title{Weighted Least Squares}
\usage{
fitWLS(y, X, w)
}
\arguments{
\item{y}{Nx1 Response vector.}

\item{X}{NxP Design matrix.}

\item{w}{Nx1 Weight vector.}
}
\value{
List containing the following:
\item{Beta}{Regression coefficient.}
\item{V}{Outcome variance.}
\item{Ibb}{Information matrix for beta.}
\item{Resid}{Outcome residuals.}
}
\description{
Fits the following weighted least squares model: 
\eqn{y_{i}=x_{i}'\beta+\epsilon_{i}}. Here, the subject-specific residual is
normally distributed with mean zero and variance \eqn{\sigma^{2}/w_{i}}.
\eqn{w_{i}} is a known, subject-specific weight, and \eqn{\sigma} is a
common scale parameter.
}
