% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LogSumExp}
\alias{LogSumExp}
\title{Log Sum Exp}
\usage{
LogSumExp(x, cum = FALSE)
}
\arguments{
\item{x}{Jx1 numeric vector.}

\item{cum}{Return the cumulative sum vector?}
}
\value{
Jx1 cumulative log sum exp vector.
}
\description{
Calculates the Jx1 log sum exp vector.
}
