% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageHelp.R
\docType{package}
\name{MatrixOps}
\alias{MatrixOps}
\alias{MatrixOps-package}
\title{MatrixOps: Matrix Operations}
\description{
Package to store matrix operations implemented in RcppArmadillo.
}
\author{
Zachary R. McCaw
}
