% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{norm}
\alias{norm}
\title{Lp Norm}
\usage{
norm(x, p = 2L)
}
\arguments{
\item{x}{Jx1 numeric vector.}

\item{p}{scalar power.}
}
\value{
Scalar norm.
}
\description{
Calculates the Lp norm of a vector
}
